Remove parser.cpp/h and wordlist generation
